+++ src/main/java/com/example/InventoryManagementSystem/dto/StockMovementResponse.java
@@ -21,7 +21,7 @@ public class StockMovementResponse {
 
     private Integer quantity;
 
-    private Integer referenceId;
+    private Long referenceId;
 
     private String notes;
